+++ contracts/mocks/MockOrderBook.sol
@@ -19,7 +19,7 @@ contract MockOrderBook is OrderBook {
         address factory_
     ) OrderBook(orderToken_, quoteToken_, baseToken_, quoteUnit_, makerFee_, takerFee_, factory_) {}
 
-    function indexToPrice(uint16 priceIndex) public pure override returns (uint128) {
+    function indexToPrice(uint16 priceIndex) public pure override returns (uint256) {
         require(priceIndex <= _MAX_INDEX, "MAX_INDEX");
         return priceIndex * _PRICE_PRECISION;
     }
